feature : 로깅 옵션 추가, launch 파일 배포 추가,route_msgs msg 현행화

+++ path_storage/path_storage/entity/node.py
@@ -10,6 +10,6 @@ class Node(BaseModel):
     position: Position
     type: Optional[str] = None
     kind: Optional[str] = None
-    heading: Optional[int] = None
+    heading: Optional[float] = None
     direction: Optional[str] = None
     detectionRange: Optional[List[DetectionRange]] = None
